requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
google-generativeai>=0.8.0
lxml>=5.0.0
importlib-metadata==4.13.0
torch>=2.0.0
nltk>=3.8.0
numpy>=1.24.0

